Revert inert/non-defect changes from static-analysis fix pass
A second independent re-verification against origin/master found that a
subset of the applied edits do not fix a reachable defect. Reverted them
to keep the diff scoped to genuine bugs; no confirmed fix is touched.

- 82/83 adapt_grid.cpp: (bigint) casts on int*sizeof(...) are inert
  (sizeof is size_t, multiply already 64-bit) -- same pattern rejected
  as bug 93. Bug 81 (nglocalprev) kept.
- 61 surf.cpp: (size_t)(nmax-old)*sizeof(Line/Tri) memset casts already
  64-bit; reverted. Bug 33 snprintf(estyle) kept.
- 60 surf_comm.cpp: spread_own2local_reduce (bigint)nlocal*n create-casts
  are inert (function already errors on bcount>MAXSMALLINT; create
  truncates to int). spread_local2own casts (genuine UB removal) kept.
- 59 surf_custom.cpp: 1-D vector memset casts inert; reverted. 2-D array
  casts (n*eicol is int*int, can overflow) kept.
- 100 compute_dt_grid.cpp + KOKKOS: vrm_max>0 guard unreachable -- loop
  already does if(!(temp[i]>0.)) continue upstream.
- 105 collide_vss.cpp: volume>0 guards unreachable -- callers hard-error
  on volume==0. Bug 46 (vremax==0) and 47 (rotc2) kept; EPSZERO intact.
- 75 react_bird_kokkos.cpp: pool-seed 12345->54321 is speculative and
  changes RNG reproducibility; reverted.

make serial links spa_serial cleanly.

+++ src/adapt_grid.cpp
@@ -984,7 +984,7 @@ void AdaptGrid::candidates_coarsen()
       if (clhash->find(parentID) == clhash->end()) {
         if (cnum == cnummax) {
           cnummax += DELTA_LIST;
-          clist = (CList *) memory->srealloc(clist,(bigint)cnummax*sizeof(CList),
+          clist = (CList *) memory->srealloc(clist,cnummax*sizeof(CList),
                                              "adapt_grid:clist");
         }
         (*clhash)[parentID] = cnum;
@@ -1050,7 +1050,7 @@ void AdaptGrid::candidates_coarsen()
       (*clhash)[parentID] = cnum;
       if (cnum == cnummax) {
         cnummax += DELTA_LIST;
-        clist = (CList *) memory->srealloc(clist,(bigint)cnummax*sizeof(CList),
+        clist = (CList *) memory->srealloc(clist,cnummax*sizeof(CList),
                                            "adapt_grid:clist");
       }
       clist[cnum].parentID = parentID;
@@ -1376,7 +1376,7 @@ void AdaptGrid::particle_surf_comm()
 
   nsend = nreturn;
   memory->create(proclist,nsend,"adapt_grid:proclist");
-  SendAdapt *sadapt = (SendAdapt *) memory->smalloc((bigint)nsend*sizeof(SendAdapt),
+  SendAdapt *sadapt = (SendAdapt *) memory->smalloc(nsend*sizeof(SendAdapt),
                                                     "adapt_grid:sadapt");
   for (int i = 0; i < nreturn; i++) {
     icell = outbuf[i].icell;
@@ -1439,7 +1439,7 @@ void AdaptGrid::particle_surf_comm()
     if (alhash->find(parentID) == alhash->end()) {
       if (anum == anummax) {
         anummax += DELTA_LIST;
-        alist = (ActionList *) memory->srealloc(alist,(bigint)anummax*sizeof(ActionList),
+        alist = (ActionList *) memory->srealloc(alist,anummax*sizeof(ActionList),
                                                 "adapt_grid:alist");
       }
       (*alhash)[parentID] = anum;
