+++ src/surf_custom.cpp
@@ -150,7 +150,7 @@ void Surf::allocate_custom(int index)
   if (etype[index] == INT) {
     if (esize[index] == 0) {
       int *ivector = memory->create(eivec[ewhich[index]],n,"surf:eivec");
-      if (ivector) memset(ivector,0,(size_t)n*sizeof(int));
+      if (ivector) memset(ivector,0,n*sizeof(int));
       eivec_local[ewhich[index]] = NULL;
     } else {
       int **iarray = memory->create(eiarray[ewhich[index]],
@@ -162,7 +162,7 @@ void Surf::allocate_custom(int index)
   } else {
     if (esize[index] == 0) {
       double *dvector = memory->create(edvec[ewhich[index]],n,"surf:edvec");
-      if (dvector) memset(dvector,0,(size_t)n*sizeof(double));
+      if (dvector) memset(dvector,0,n*sizeof(double));
       edvec_local[ewhich[index]] = NULL;
     } else {
       double **darray = memory->create(edarray[ewhich[index]],
@@ -193,7 +193,7 @@ void Surf::reallocate_custom()
     if (etype[index] == INT) {
       if (esize[index] == 0) {
         int *ivector = memory->grow(eivec[ewhich[index]],nnew,"surf:eivec");
-        if (nnew > nold) memset(&ivector[nold],0,(size_t)(nnew-nold)*sizeof(int));
+        if (nnew > nold) memset(&ivector[nold],0,(nnew-nold)*sizeof(int));
       } else {
         int **iarray = memory->grow(eiarray[ewhich[index]],
                                     nnew,eicol[ewhich[index]],"surf:eiarray");
@@ -204,7 +204,7 @@ void Surf::reallocate_custom()
     } else {
       if (esize[index] == 0) {
         double *dvector = memory->grow(edvec[ewhich[index]],nnew,"surf:edvec");
-        if (nnew > nold) memset(&dvector[nold],0,(size_t)(nnew-nold)*sizeof(double));
+        if (nnew > nold) memset(&dvector[nold],0,(nnew-nold)*sizeof(double));
       } else {
         double **darray = memory->grow(edarray[ewhich[index]],
                                        nnew,edcol[ewhich[index]],"surf:eearray");
